refactor: to get clean api for block states
